Reduce references/Inc. readability + "Async" fix
When using the fetchInfo function, there is no pause in the script unlike the fetchJSON function which runs in async mode. This causes the two functions that serve the same purpose to act differently due to how they end up letting Danbooru's JavaScript run. Using a 1 millisecond timer to interrupt the flow of this script when using fetchInfo is enough to fix this.