Simple layout option #1
Trying to find relatively simple ways to refine the layout. Here's a rough first concept.

* Create several arrays denoting sections that try to split up the settings into certain categories. These arrays also determine the order of the settings for each section.
* Create a simple function that will use the arrays to make each section. It'll append a custom header/title to each section and create each setting in the same fashion already employed.

A similar option to this would not use arrays and rely on a function that would create one (or more settings) at a time. The other portions of the layout would be created manually and the function would be called to insert one or more settings where desired.