Hiddent comment fixes + Tag list scrollbars
* Fixed the problems with missing info on hidden images in the comments section.
* Added an option for tweaking the tag lists on individual posts with a limited height and scrollbars.

The scrollbars are a style I've been using in stylish for a long time and basically work the same way as the scrolling "favorited by" list that used to be in the script. It'll allow the user to easily limit the length of the tag lists so that they don't move all the sidebar options below them way off the screen. I hope I didn't word the description for it too oddly...