Cleanup
For the average user that just wants access to the occasional hidden image, this is pretty much a good alternative. For the user that outright searches for the hidden tags, the alternative isn't that great.

What I've observed:
1) Hidden images get their information removed only if they're at the start of the results.
2) If you have a hidden image anywhere after a regular image in the results, it's information is not removed.
3) Banned and deleted posts do not count as actual hidden images and retain all their info.

In short, if you search for the hidden tags, every post will have it's information removed. If you're just doing an average search, any consecutive hidden images at the start of the results will have their info removed while all other hidden images that come after a regular image will display properly.

The errors (Failbooru) were making things quite a pain to test... Considering other approaches, but this might be good as a starting release.